Refactor WithBadElements
