+++ packages/alfa-rules/src/common/act/diagnostic.ts
@@ -7,18 +7,20 @@ import { WithFirstHeading } from "../../sia-r61/rule";
 import { DistinguishingStyles } from "../../sia-r62/diagnostics";
 import { DistinguishingStyles as DeprecatedDistinguishingStyles } from "../../sia-dr62/rule";
 import { MatchingClasses } from "../../sia-r65/diagnostics";
-import { DeprecatedElements } from "../../sia-r70/rule";
 import { WithDeclaration } from "../../sia-r75/rule";
 import { WithNextHeading } from "../../sia-r78/rule";
 import { ClippingAncestors } from "../../sia-r83/rule";
 import { Languages } from "../../sia-r109/rule";
 
+// R42, R60, R68, R76
+import { WithRole } from "../diagnostic/with-role";
+
 // R66, R69
 import { Contrast } from "../diagnostic/contrast";
 import { ColorError } from "../dom/get-colors";
 
-// R42, R60, R68, R76
-import { WithRole } from "../diagnostic/with-role";
+// R70, R90
+import { WithBadElements } from "../diagnostic/with-bad-elements";
 
 /**
  * @public
@@ -35,8 +37,6 @@ export namespace Diagnostic {
 
   export const { isContrast } = Contrast;
 
-  export const { isDeprecatedElements } = DeprecatedElements;
-
   export const { isDistinguishingStyles } = DistinguishingStyles;
 
   export const { isDistinguishingStyles: isDistinguishingStylesDeprecated } =
@@ -51,6 +51,8 @@ export namespace Diagnostic {
   export const { isRoleAndRequiredAttributes } = RoleAndRequiredAttributes;
 
   export const { isSameNames } = SameNames;
+
+  export const { isWithBadElements } = WithBadElements;
 
   export const { isWithDeclaration } = WithDeclaration;
